Tighten MessageText click dispatch and add snapshot coverage
Apply audit findings from the gesture review:

- handleAnnotationClick now filters firstOrNull by isInteractiveTag.
  Previously, when a non-interactive annotation (e.g. a custom
  decoration added through AnnotatedMessageTextBuilder) overlapped
  a URL/email/mention range, the first-overlapping non-interactive
  one was returned and the when fell through silently. The link
  never opened. Locked in by a new MessageTextHelpersTest case
  covering the overlap.

- ClickableText's pointerInput now keys on Unit and reads its
  callbacks through rememberUpdatedState. The block is no longer
  cancelled and restarted each composition because the caller
  allocates fresh lambdas.

- styledText.getStringAnnotations is wrapped in remember(styledText)
  so the list is not reallocated on every recomposition.

- A small WHY comment is added on the two non-obvious gesture
  decisions in ClickableText (the early-return for non-interactive
  characters, and consumeUntilUp after long-press).

Tests:
- The previous MessageTextTest is renamed to MessageTextHelpersTest
  to free the canonical name for the new snapshot test, matching
  the codebase convention (e.g. ReactionsMenuContentTest).
- MessageTextTest is the new Paparazzi snapshot suite. It covers
  plain text, URL, email, mention, and URL + mention scenarios.
  The fixtures live next to the production code as internal
  preview-friendly composables (MessageTextPlain, MessageTextWithUrl,
  ...) so the @Preview composables and the snapshot tests share
  the same definitions.

The earlier audit also recommended Compose UI gesture tests
(tap / long-press / drag-out). Those are not included: the
gesture loop relies on withTimeout inside awaitPointerEventScope,
which the Robolectric test environment does not drive reliably
for this case. The behaviour stays under manual QA.
